\name{blim-package}
\alias{blim-package}
\docType{package}
\title{
\packageTitle{blim}
}
\description{
\packageDescription{blim}
}


\author{
\packageAuthor{blim}

Maintainer: \packageMaintainer{blim}
}
\references{
Dynamic tuning procedure using information from
https://support.sas.com/documentation/cdl/en/statug/63033/HTML/default/viewer.htm#statug_mcmc_sect022.htm

Beland, S., Klugkist, I., Raiche, G. & Magis, D. (2012). A short introduction 
into Bayesian evaluation of informative hypotheses as an alternative to 
exploratory comparisons of multiple group means. Tutorials in Quantitative 
Methods for Psychology 8(2), p. 122-126.

Raftery, A. E. (1996). Hypothesis testing and model selection. Markov chain 
Monte Carlo in practice, 163-188.
}

\keyword{ package, blim, bayesian, linear, modelling, univariate }

\seealso{

\code{\link[<blim.Rd>]{blim}}

}
\examples{

}
