% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimate_Blim.R
\name{Blim_estim}
\alias{Blim_estim}
\alias{Blim_estim.default}
\alias{Blim_estim.sam}
\alias{Blim_estim.FLStock}
\title{Function to estimate Blim given a time series of recruitment and SSB}
\usage{
Blim_estim(data, ...)

\method{Blim_estim}{default}(
  data,
  Rage,
  years = as.numeric(rownames(data)),
  is.spasmodic = FALSE,
  types = 2:6,
  doplot = FALSE,
  spasmodic_prob = 0.9,
  narrow_threshold = 0.25,
  conf_level_lm = 0.95,
  ...
)

\method{Blim_estim}{sam}(
  data,
  years = as.numeric(rownames(rectable(data))),
  is.spasmodic = FALSE,
  types = 2:6,
  doplot = FALSE,
  spasmodic_prob = 0.9,
  narrow_threshold = 0.25,
  conf_level_lm = 0.95,
  ...
)

\method{Blim_estim}{FLStock}(
  data,
  years = as.numeric(names(rec(data)[drop = TRUE])),
  is.spasmodic = FALSE,
  types = 2:6,
  doplot = FALSE,
  spasmodic_prob = 0.9,
  narrow_threshold = 0.25,
  conf_level_lm = 0.95,
  ...
)
}
\arguments{
\item{data}{SAM fit, FLStock object, or data frame with columns "Rec" and "SSB" and with years as row names.}

\item{...}{extra arguments for internal functions.}

\item{Rage}{age at recruitment to correctly identify SSB-Rec pairs, only needed if data is a data frame.}

\item{years}{the years to consider for the SSB-Rec pairs, default is the whole time series.}

\item{is.spasmodic}{\code{TRUE} if the stock is assumed to have spasmodic recruitment, default is \code{FALSE}.}

\item{types}{stock types to consider for Blim estimation, default are types 2 to 6 in ICES (2021).}

\item{doplot}{if TRUE, the final stock-recruitment plot will be drawn during estimation, default is \code{FALSE}.}

\item{spasmodic_prob}{the probability to use to identify outliers in recruitment in the quantile function when \code{is.spasmodic=TRUE}, default is 0.9.}

\item{conf_level_lm}{confidence level of the confidence interval of the slope parameter of the linear regression (types 3 to 6), default is 0.95.}

\item{narrow_treshold}{threshold to determine if the range of SSB is narrow or not (difference between stock types 5 and 6), default is 0.25. The range of SSB is considered narrow if less than \code{narrow_threshold} of SSB is below max(SSB)/2.}
}
\value{
Blim and stock type code as attribute.
}
\description{
Function to estimate Blim given a time series of recruitment and SSB
}
\details{
The function estimates Blim according to the different stock types in ICES (2021, \url{https://doi.org/10.17895/ices.advice.7891}).
If \code{is.spasmodic=TRUE}, stock type 1, Blim is estimated as the lowest SSB, where large recruitment (probability \code{>=spasmodic_prob}) is observed, default is \code{FALSE}. Stock type 1 will only be considered if the argument is \code{TRUE}. \cr
\cr 
By default, the function will estimate Blim according to the stock types 2-6.
First, the function tries to estimate Blim as the inflection point of a hockey-stick stock-recruitment (SRR) curve (stock type 2). 
If the inflection point is outside the range of observed SSB, the stock type is considered to be either 5 or 6 if the inflection point 
is below the minimum SSB, or type 3 if above. 
If a consistent SRR cannot be estimated (non convergence), Blim will be estimated according to types 3-6 as explained below. \cr
\cr
To determine which type is the best between types 3-6, an unconstrained linear regression is fitted to the SSB-Rec pairs.
The slope indicates which type the stock is and Blim is estimated accordingly to ICES (2021).
If the slope of the linear regression is not different from zero (confidence interval given \code{conf_level_lm} contains zero), 
the SRR is considered to be flat (types 5 or 6) and the final type is chosen depending on the range of SSB values (narrow or not). If the slope is positive, the type 3 is assumed and the type 4 is assumed otherwise.                                         
\cr
It is possible to avoid fitting a hockey-stick SRR by using \code{types!=2}. In this case, any stock type between 3 and 6 can be considered to estimate Blim.
}
\examples{
s <- seq(100, 2000, length=50)
r <- exp((2+log(s)-0.001*s)+rnorm(length(s)))
data <- data.frame(Rec=r, SSB=s, row.names=1971:2020)

# The stock has spasmodic recruitment (type 1):
Blim_estim(data, Rage=0, is.spasmodic=TRUE, doplot=TRUE) # only type 1 is assumed

# The algorithm tests for ICES stock types 2-6 and chooses the best:
Blim_estim(data, Rage=0, doplot=TRUE) 

# Type 2 is not considered and the algorithm only tests for types 3-6:
Blim_estim(data, Rage=0, types=3:6, doplot=TRUE) 
}
\author{
Vanessa Trijoulet
}
