\name{blim-package}
\alias{blim-package}
\docType{package}
\title{
\packageTitle{blim}
}
\description{
\packageDescription{blim}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{blim}
\packageIndices{blim}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{blim}

Maintainer: \packageMaintainer{blim}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
